Refactor Firebird utils into separate modules
