+++ utils/db/firebird_manager.py
@@ -18,11 +18,15 @@ from concurrent.futures import ThreadPoolExecutor
 from typing import Dict, List, Optional, Any, Set, AsyncGenerator, Tuple
 from dataclasses import dataclass, field
 from datetime import datetime, date
-from enum import IntEnum
 from contextlib import contextmanager, asynccontextmanager
-import re
 from models.container_event import TrackingResult
 from utils.logging import get_logger
+
+from .connection import FirebirdConnectionManager
+from .transformer import FirebirdDateTransformer
+from .matcher import FirebirdOperationMatcher
+from .stats import FirebirdStatisticsCollector
+from .models import EntityStatusID, EntityColumnMapping, EntityTableConfig
 
 try:
     import firebird.driver as fdb  # pip install firebird-driver
@@ -31,240 +35,6 @@ except ImportError:
     FIREBIRD_AVAILABLE = False
     fdb = None
 
-
-# =============================================================================
-# КОНФИГУРАЦИОННЫЕ МОДЕЛИ - твоя идея с dataclasses!
-# =============================================================================
-
-class EntityStatusID(IntEnum):
-    """
-    Важные статусы entity как Enum - типобезопасно и читаемо
-    Основано на твоем JSON файле tables
-    """
-    NEW = 1 # "00. Новый"
-    LOCATION_RECEIVED = 2 # "01. Место получено"
-    SEA = 3 # "03. Море"
-    ARRIVED_AT_STATION = 4 # "07. Прибыл на станцию"
-    TERMINAL_OPERATION = 5 # "04. ТО"
-    WAITING_DEPARTURE = 6 # "02. Жду выход"
-    RAILWAY = 7 # "06. ЖД"
-    TRANSPORTATION_CLOSED = 8 # "09. Перевозка закрыта"
-    DELIVERED = 9 # "08. Доставлен"
-    DOCUMENTS_RECEIVED = 12 # "08..Документы получены"
-    ATTENTION = 13 # "08...Внимание"
-    WAITING_SHIPMENT = 15 # "05. Ожидает отгрузку"
-    RAID = 16 # "03. Рейд"
-    DIRECT_CAR = 17 # "03. Прямое авто"
-    RAIL = 23 #  "03. RAIL"
-    CANCELLED = 24  # "99. Отмена"
-    UNLOADING = 25 # "03..Выгрузка"
-    CLIENT_DEBT = 26 # "08..Долг Клиента"
-    OUTPUT = 27 # "04..Выпуск"
-    LCL_RU = 28 # "06. LCL по РФ"
-    NO_AVP = 29 # "08..нет АВП"
-    RAILWAY_OUTPUT = 30 # "07..Выпуск ЖД"
-    PP_DIRECT_RAILWAY = 31 # "05.. П\/П прямое ЖД"
-    LOCAL_ISSUANCE = 32 # "04..Выпуск МВ"
-    AWAITS_AVAILABILITY = 33 # "00.1 Ожидает готовности"
-    LOOKING_PLACE = 34 # "00.2 Ищем место"
-    WAITING_PLACE = 35 # "00.3 Ждем место"
-    TRANSHIPMET = 36 # "03.Трансшипмент"
-    GIVEN_FOR_CLOSE = 37 # "08..Передана на закрытие"
-    SEA_THROUGH = 41 # "03. Море \/ СКВОЗНОЙ СЕРВИС"
-    RAID_THROUGH = 42 # "03. Рейд\/ СКВОЗНОЙ СЕРВИС"
-    UNLOADING_THROUGH = 43 # "03..Выгрузка\/ СКВОЗНОЙ СЕРВИС"
-    TERMINAL_OPERATION_THROUGH = 44 # "04. ТО\/ СКВОЗНОЙ СЕРВИС"
-    RELEASE_THROUGH = 45 # "04..Выпуск\/ СКВОЗНОЙ СЕРВИС"
-    WAITING_SHIPMENT_THROUGH = 46 # "05. Ожидает отгрузку\/ СКВОЗНОЙ СЕРВИС"
-    WAITING_EMPTY = 52 # "08..Ожидаем сдачу порожнего"
-    AUTO_CN = 55 # "03. Авто Китай"
-    AUTO_BORDER_CROSSING = 56 # "03..Переход границы (авто)"
-    TERMINAL_OPERATION_AUTO = 57 # "ТО (прямое авто)"
-    FTL_RELEASE = 58 # "04...Выпуск FTL"
-
-    
-    @classmethod
-    def get_excluded_statuses(cls) -> Set[int]:
-        """Статусы, которые исключаем из обработки"""
-        return {
-            int(cls.TRANSPORTATION_CLOSED),  # 8 - "09. Перевозка закрыта"
-            int(cls.DELIVERED),              # 9 - "08. Доставлен" 
-           int(cls.CANCELLED)               # 24 - "99. Отмена"
-        }
-
-
-@dataclass(frozen=True)
-class EntityColumnMapping:
-    """
-    Маппинг колонки entity на операцию FESCO
-    Использует твою концепцию ColumnMapping, но адаптированную под entity
-    """
-    entity_column: str                  # Колонка в entity (например, DATE_ETA)
-    fesco_field: str                    # Поле из FESCO API
-    operation_patterns: Tuple[str, ...] # Паттерны операций для этой колонки
-    transform_func: str = "firebird_date"
-    priority: int = 0                   # Приоритет при совпадении нескольких паттернов
-    description: str = ""               # Описание для документации
-    column_datatype: str = "DATE"       # Тип данных колонки: "DATE", "TIMESTAMP", "INTEGER"
-    
-
-    def __post_init__(self):
-        if not self.operation_patterns:
-            raise ValueError("operation_patterns должен cодержать хотя бы один паттерн")
-
-    def get_transform_datatype(self) -> str:
-        """Определяет стратегию трансформации на основе типа колонки"""
-        if self.column_datatype == "TIMESTAMP":
-            return "firebird_timestamp"
-        elif self.column_datatype == "DATE":
-            return "firebird_date_only"
-        elif self.column_datatype == "INTEGER":
-            return "firebird_integer"
-        else:
-            return "firebird_date"  # fallback
-
-    def matches_operation(self, operation: str) -> float:
-        """
-        Проверяет соответствие операции маппингу и возвращает оценку
-        
-        Returns:
-            float: Оценка от 0.0 до 1.0 (нормализованная)
-        """
-        if not operation:
-            return 0.0
-            
-        operation_lower = operation.lower().strip()
-        max_score = 0.0
-        
-        for pattern in self.operation_patterns:
-            pattern_lower = pattern.lower().strip()
-            
-            # Точное совпадение
-            if pattern_lower == operation_lower:
-                max_score = max(max_score, 1.0)
-            # Операция содержит паттерн
-            elif pattern_lower in operation_lower:
-                coverage = len(pattern_lower) / len(operation_lower)
-                position_bonus = 0.2 if operation_lower.startswith(pattern_lower) else 0.1
-                score = 0.8 * coverage + position_bonus
-                max_score = max(max_score, score)
-            # Паттерн содержит операцию  
-            elif operation_lower in pattern_lower:
-                score = 0.6 * len(operation_lower) / len(pattern_lower)
-                max_score = max(max_score, score)
-        
-        # Нормализуем с учетом приоритета (но не превышаем 1.0)
-        priority_bonus = min(0.1, self.priority / 100)
-        return min(1.0, max_score + priority_bonus)
-
-
-@dataclass
-class EntityTableConfig:
-    """
-    Конфигурация entity таблицы
-    """
-    # Основные колонки таблицы
-    table_name: str = "ENTITY"  # Firebird обычно в верхнем регистре
-    primary_key: str = "ID"
-    container_column: str = "NAME"
-    status_column: str = "SP_ENTITY_STATUS_ID"
-    line_column: str = "LEGAL_PERSON_LINE_ID"
-    
-    # Колонки дат
-    date_eta: str = "DATE_ETA"
-    date_etd: str = "DATE_ETD"
-    date_in: str = "DATE_IN"
-    date_railway_loading: str = "DATE_RAILWAY_LOADING"
-    date_railway_delivery: str = "DATE_RAILWAY_DELIVERY"
-    remaining_distance: str = "TRACING_DAYS"
-    
-    # Маппинги операций FESCO на колонки дат
-    date_mappings: Dict[str, EntityColumnMapping] = field(default_factory=dict)
-    
-    # Статусы для исключения из обработки
-    excluded_status_ids: Set[int] = field(default_factory=lambda: EntityStatusID.get_excluded_statuses())
-    
-    def __post_init__(self):
-        """Создаем маппинги по умолчанию если не указаны"""
-        if not self.date_mappings:
-            self.date_mappings = self._create_default_mappings()
-        self._validate_config()
-
-        self.excluded_status_ids = {int(v) for v in self.excluded_status_ids}
-
-    def _validate_config(self):
-        """Валидация конфигурации"""
-        if not self.table_name or not self.table_name.strip():
-            raise ValueError("table_name не может быть пустым")
-        
-        required_columns = [
-            self.primary_key, self.container_column, 
-            self.status_column, self.line_column
-        ]
-        
-        for column in required_columns:
-            if not column or not column.strip():
-                raise ValueError(f"Обязательная колонка не может быть пустой: {column}")
-    
-    def _create_default_mappings(self) -> Dict[str, EntityColumnMapping]:
-        """
-        Создает маппинги операций FESCO на колонки
-        """
-        return {
-            self.date_eta: EntityColumnMapping(
-                entity_column=self.date_eta,
-                fesco_field="date",
-                operation_patterns=["Выгружается груженным"], # type: ignore
-                priority=10,
-                description="Estimated Time of Arrival",
-                column_datatype="DATE"
-            ),
-            
-            self.date_etd: EntityColumnMapping(
-                entity_column=self.date_etd,
-                fesco_field="date",
-                operation_patterns=["Грузится на фидер", "Loading Feeder Full"], # type: ignore
-                priority=10,
-                description="Estimated Time of Departure",
-                column_datatype="DATE"
-            ),
-            
-            self.date_in: EntityColumnMapping(
-                entity_column=self.date_in,
-                fesco_field="date",
-                operation_patterns=["Прием с моря", "Регистрация ДО1", "Discharged from vessel", "DO1 registration"], # type: ignore
-                priority=8,
-                description="Выгрузка на терминал",
-                column_datatype="TIMESTAMP"
-            ),
-            
-            self.date_railway_loading: EntityColumnMapping(
-                entity_column=self.date_railway_loading,
-                fesco_field="date",
-                operation_patterns=["Отправление вагона со станции", "Wagon has left the station"], # type: ignore
-                priority=8,
-                description="Отгрузка на платформу",
-                column_datatype="DATE"
-            ),
-            
-            self.date_railway_delivery: EntityColumnMapping(
-                entity_column=self.date_railway_delivery,
-                fesco_field="date",
-                operation_patterns=["Документы для отправки по ЖД приняты", "Documents for sending by railway accepted"], # type: ignore
-                priority=8,
-                description="Сдача на ж/д",
-                column_datatype="DATE"
-            ),
-
-            self.remaining_distance: EntityColumnMapping(
-                entity_column=self.remaining_distance,
-                fesco_field="remainingDistance",
-                operation_patterns=("",),
-                priority=10,
-                description="Слежение",
-                column_datatype="INTEGER"
-            )
-        }
 
 
 @dataclass
@@ -292,410 +62,6 @@ class ContainerInfo:
 # =============================================================================
 # ГЛАВНЫЙ КЛАСС - ЕДИНЫЙ МЕНЕДЖЕР ENTITY
 # =============================================================================
-
-class FirebirdConnectionManager:
-    
-    def __init__(self, firebird_config: dict, entity_config: EntityTableConfig = None):
-        if not FIREBIRD_AVAILABLE:
-            raise ImportError(
-                "❌ Firebird драйвер не установлен!\n"
-                "Установите: pip install fdb\n"
-                "или: pip install firebird-driver"
-            )
-        
-        self.config = firebird_config
-        self.logger = get_logger("firebird.connection")
-        self._validate_config()
-        
-        # Thread-safe счетчик соединений
-        self._connection_lock = threading.Lock()
-        self._active_connections = 0
-        self._max_connections = 10
-    
-    def _validate_config(self):
-        """Валидация конфигурации подключения"""
-        required_fields = ['host', 'database', 'user', 'password']
-        for field in required_fields:
-            if not self.config.get(field):
-                raise ValueError(f"Отсутствует обязательное поле: {field}")
-    
-    @contextmanager
-    def get_connection(self):
-        """
-        Context manager для получения подключения
-        
-        Гарантирует правильное закрытие и учет активных соединений
-        """
-        connection = None
-        
-        # Проверяем лимит соединений
-        with self._connection_lock:
-            if self._active_connections >= self._max_connections:
-                raise RuntimeError(f"Превышен лимит соединений: {self._max_connections}")
-            self._active_connections += 1
-        
-        try:
-            connection = self._create_connection()
-            yield connection
-            
-        except Exception as e:
-            self.logger.error(f"❌ Ошибка работы с подключением: {e}")
-            raise
-            
-        finally:
-            # Всегда освобождаем ресурсы
-            if connection:
-                try:
-                    connection.close()
-                except Exception as close_error:
-                    self.logger.error(f"Ошибка закрытия соединения: {close_error}")
-            
-            with self._connection_lock:
-                self._active_connections -= 1
-    
-    def _create_connection(self):
-        """Создание нового подключения к Firebird"""
-        try:
-            if 'dsn' in self.config:
-                dsn = self.config['dsn']
-            else:
-                host = self.config['host']
-                database = self.config['database']
-                dsn = f"{host}:{database}"
-            
-            connection = fdb.connect(
-                database=self.config['database'],
-                user=self.config['user'],
-                password=self.config['password'],
-                charset='UTF8'
-            )
-            
-            return connection
-            
-        except Exception as e:
-            self.logger.error(f"❌ Ошибка создания Firebird подключения: {e}")
-            raise
-    
-    async def test_connection(self) -> bool:
-        """Тестирование подключения"""
-        try:
-            with self.get_connection() as conn:
-                cursor = conn.cursor()
-                cursor.execute("SELECT 1 FROM RDB$DATABASE")
-                result = cursor.fetchone()
-                return result is not None
-                
-        except Exception as e:
-            self.logger.error(f"❌ Ошибка тестирования подключения: {e}")
-            return False
-
-class FirebirdDateTransformer:
-    """
-    Трансформация дат из FESCO в форматы Firebird
-    
-    Отвечает ТОЛЬКО за преобразование типов данных
-    """
-    
-    def __init__(self):
-        self.logger = get_logger("firebird.transformer")
-        
-        # Предкомпилированные regex для производительности
-        self._number_pattern = re.compile(r'\d+')
-        
-        # Поддерживаемые форматы дат FESCO
-        self._timestamp_formats = [
-            "%Y-%m-%d %H:%M:%S",
-            "%Y-%m-%dT%H:%M:%S",
-            "%d.%m.%Y %H:%M:%S",
-            "%d.%m.%Y %H:%M",
-        ]
-        
-        self._date_formats = [
-            "%Y-%m-%d",
-            "%d.%m.%Y",
-            "%Y/%m/%d",
-            "%d/%m/%Y"
-        ]
-    
-    def transform_value(self, value: Any, target_type: str) -> Optional[Any]:
-        """
-        Универсальный метод трансформации значений
-        
-        Args:
-            value: Исходное значение от FESCO
-            target_type: Целевой тип ("DATE", "TIMESTAMP", "INTEGER")
-            
-        Returns:
-            Преобразованное значение или None при ошибке
-        """
-        if value is None:
-            return None
-        
-        try:
-            if target_type == "TIMESTAMP":
-                return self._transform_to_timestamp(value)
-            elif target_type == "DATE":
-                return self._transform_to_date(value)
-            elif target_type == "INTEGER":
-                return self._transform_to_integer(value)
-            else:
-                self.logger.warning(f"Неизвестный тип: {target_type}")
-                return str(value)
-                
-        except Exception as e:
-            self.logger.warning(f"Ошибка трансформации '{value}' в {target_type}: {e}")
-            return None
-    
-    def _transform_to_timestamp(self, date_str: str) -> Optional[datetime]:
-        """Трансформация в TIMESTAMP (дата + время)"""
-        if not date_str or not str(date_str).strip():
-            return None
-        
-        cleaned = str(date_str).strip()
-        
-        # Пробуем форматы с временем
-        for fmt in self._timestamp_formats:
-            try:
-                return datetime.strptime(cleaned, fmt)
-            except ValueError:
-                continue
-        
-        # Если время не указано, пробуем форматы даты + добавляем время 00:00:00
-        for fmt in self._date_formats:
-            try:
-                date_only = datetime.strptime(cleaned, fmt)
-                return date_only  # datetime автоматически добавит 00:00:00
-            except ValueError:
-                continue
-        
-        return None
-    
-    def _transform_to_date(self, date_str: str) -> Optional[date]:
-        """Трансформация в DATE (только дата)"""
-        timestamp = self._transform_to_timestamp(date_str)
-        return timestamp.date() if timestamp else None
-    
-    def _transform_to_integer(self, value_str: str) -> Optional[int]:
-        """Трансформация в INTEGER (извлечение числа)"""
-        if not value_str:
-            return None
-        
-        # Извлекаем первое число из строки
-        numbers = self._number_pattern.findall(str(value_str))
-        if numbers:
-            try:
-                return int(numbers[0])
-            except ValueError:
-                pass
-        
-        return None
-
-
-class FirebirdOperationMatcher:
-    """
-    Сопоставление операций FESCO с колонками entity
-    
-    Отвечает ТОЛЬКО за логику маппинга операций
-    """
-    
-    def __init__(self, entity_config: EntityTableConfig):
-        self.entity_config = entity_config
-        self.logger = get_logger("firebird.matcher")
-    
-    def find_best_mapping(self, operation: str) -> Optional[EntityColumnMapping]:
-        """
-        Найти лучший маппинг для операции FESCO
-        
-        Returns:
-            Лучший маппинг или None если не найден
-        """
-        if not operation or not operation.strip():
-            return None
-        
-        operation_clean = operation.strip()
-        self.logger.debug(f"🎯 Анализируем операцию: '{operation_clean}'")
-        
-        # Собираем все маппинги с оценками
-        scored_mappings = []
-        
-        for column_name, mapping in self.entity_config.date_mappings.items():
-            score = mapping.matches_operation(operation_clean)
-            
-            if score > 0:
-                scored_mappings.append((mapping, score))
-                self.logger.debug(f"  📊 {column_name}: score={score:.3f}")
-        
-        if not scored_mappings:
-            self.logger.debug("  🤷 Подходящих маппингов не найдено")
-            return None
-        
-        # Сортируем по оценке (лучшие первыми)
-        scored_mappings.sort(key=lambda x: x[1], reverse=True)
-        
-        best_mapping, best_score = scored_mappings[0]
-        
-        # Проверяем минимальный порог качества
-        if best_score < 0.3:  # Настраиваемый порог
-            self.logger.debug(f"🚫 Лучшая оценка {best_score:.3f} ниже порога 0.3")
-            return None
-        
-        self.logger.debug(f"🎯 Выбран маппинг: {best_mapping.entity_column} (score: {best_score:.3f})")
-        return best_mapping
-    
-    def suggest_new_mappings(self, unmapped_operations: List[str]) -> Dict[str, List[str]]:
-        """
-        Предложить новые маппинги для неопознанных операций
-        
-        Returns:
-            Словарь предложений по улучшению маппингов
-        """
-        suggestions = {}
-        
-        for operation in unmapped_operations:
-            # Ищем частичные совпадения с существующими маппингами
-            for column_name, mapping in self.entity_config.date_mappings.items():
-                similarity = self._calculate_similarity(operation, mapping.operation_patterns)
-                
-                if 0.1 < similarity < 0.3:  # Похоже, но не проходит порог
-                    if column_name not in suggestions:
-                        suggestions[column_name] = []
-                    
-                    suggestions[column_name].append(
-                        f"Рассмотреть добавление паттерна для '{operation}' "
-                        f"(схожесть: {similarity:.2f})"
-                    )
-        
-        return suggestions
-    
-    def _calculate_similarity(self, operation: str, patterns: Tuple[str, ...]) -> float:
-        """Вычислить схожесть операции с паттернами"""
-        if not patterns:
-            return 0.0
-        
-        operation_words = set(operation.lower().split())
-        max_similarity = 0.0
-        
-        for pattern in patterns:
-            pattern_words = set(pattern.lower().split())
-            
-            if operation_words and pattern_words:
-                intersection = operation_words.intersection(pattern_words)
-                union = operation_words.union(pattern_words)
-                
-                if union:
-                    similarity = len(intersection) / len(union)
-                    max_similarity = max(max_similarity, similarity)
-        
-        return max_similarity
-
-
-class FirebirdStatisticsCollector:
-    """
-    Сбор и анализ статистики работы
-    
-    Отвечает ТОЛЬКО за метрики и аналитику
-    """
-    
-    def __init__(self):
-        self.stats = {
-            # Статистика чтения
-            'containers_loaded': 0,
-            'containers_filtered': 0,
-            'batches_processed': 0,
-            
-            # Статистика записи (из твоего ExternalDatabaseWriter)
-            'records_updated': 0,
-            'records_failed': 0,
-            'date_columns_updated': {},
-            'operations_processed': {},
-            
-            # Общая статистика
-            'connections_created': 0,
-            'transactions_committed': 0,
-            'transactions_rollbacked': 0,
-            'operation_times': [],
-        }
-        self.logger = get_logger("firebird.stats")
-    
-    def record_container_loaded(self, count: int = 1):
-        """Зафиксировать загрузку контейнеров"""
-        self.stats['containers_loaded'] += count
-    
-    def record_container_filtered(self, count: int = 1):
-        """Зафиксировать фильтрацию контейнеров"""
-        self.stats['containers_filtered'] += count
-    
-    def record_batch_processed(self):
-        """Зафиксировать обработку батча"""
-        self.stats['batches_processed'] += 1
-    
-    def record_update_success(self, column_name: str, operation: str):
-        """Зафиксировать успешное обновление"""
-        self.stats['records_updated'] += 1
-        
-        # Статистика по колонкам
-        if column_name not in self.stats['date_columns_updated']:
-            self.stats['date_columns_updated'][column_name] = 0
-        self.stats['date_columns_updated'][column_name] += 1
-        
-        # Статистика по операциям (ограничиваем длину)
-        operation_key = operation[:50] if operation else "EMPTY_OPERATION"
-        if operation_key not in self.stats['operations_processed']:
-            self.stats['operations_processed'][operation_key] = 0
-        self.stats['operations_processed'][operation_key] += 1
-    
-    def record_update_failure(self):
-        """Зафиксировать неудачное обновление"""
-        self.stats['records_failed'] += 1
-    
-    def record_operation_time(self, time_ms: float):
-        """Зафиксировать время выполнения операции"""
-        # Храним только последние 1000 операций для расчета среднего
-        self.stats['operation_times'].append(time_ms)
-        if len(self.stats['operation_times']) > 1000:
-            self.stats['operation_times'] = self.stats['operation_times'][-1000:]
-    
-    def get_summary(self) -> Dict[str, Any]:
-        """Получить сводную статистику"""
-        operation_times = self.stats['operation_times']
-        
-        summary = {
-            'totals': {
-                'containers_loaded': self.stats['containers_loaded'],
-                'containers_filtered': self.stats['containers_filtered'],
-                'batches_processed': self.stats['batches_processed'],
-                'records_updated': self.stats['records_updated'],
-                'records_failed': self.stats['records_failed'],
-            },
-            'success_rate': self._calculate_success_rate(),
-            'performance': {
-                'avg_operation_time_ms': sum(operation_times) / len(operation_times) if operation_times else 0,
-                'min_operation_time_ms': min(operation_times) if operation_times else 0,
-                'max_operation_time_ms': max(operation_times) if operation_times else 0,
-            },
-            'top_columns': self._get_top_columns(),
-            'top_operations': self._get_top_operations()
-        }
-        
-        return summary
-    
-    def _calculate_success_rate(self) -> float:
-        """Вычислить процент успешных операций"""
-        total = self.stats['records_updated'] + self.stats['records_failed']
-        return (self.stats['records_updated'] / total * 100) if total > 0 else 0.0
-    
-    def _get_top_columns(self, limit: int = 5) -> List[Tuple[str, int]]:
-        """Получить топ обновляемых колонок"""
-        items = list(self.stats['date_columns_updated'].items())
-        items.sort(key=lambda x: x[1], reverse=True)
-        return items[:limit]
-    
-    def _get_top_operations(self, limit: int = 5) -> List[Tuple[str, int]]:
-        """Получить топ операций"""
-        items = list(self.stats['operations_processed'].items())
-        items.sort(key=lambda x: x[1], reverse=True)
-        return items[:limit]
-
 
 # =============================================================================
 # ГЛАВНЫЙ КЛАСС - УПРОЩЕННЫЙ И НАДЕЖНЫЙ
